fix(spike): D3D11 NVENC spike — uint/int cast mismatches
Fix 6 build errors + 1 warning:
  error CS0266: Cannot implicitly convert type 'uint' to 'int'
  error CS0266: Cannot implicitly convert type 'int' to 'uint'

Root cause:
  In Vortice 3.6.2:
  - ModeDescription.Width/Height are uint
  - Texture2DDescription.Width/Height are uint
  - Our NV_ENC_REGISTER_RESOURCE struct uses int (matches NVENC SDK C header)
  - Our local int desktopW/desktopH variables need casting when assigned
    to uint fields

Fixes:
  - Phase3_TextureOwnership.cs (lines 86-87): cast ModeDescription.Width/Height
    from uint to int for local variable assignment
  - Phase4_NVENCRegistration.cs (lines 219-220): cast ModeDescription.Width/Height
    from uint to int when populating NV_ENC_REGISTER_RESOURCE.width/height
  - Phase5_PerformanceBenchmark.cs (lines 113-114): cast desktopW/desktopH
    from int to uint when populating Texture2DDescription.Width/Height
  - Phase1_DeviceTest.cs (line 54): add null-forgiving operator (factory!.)
    to silence CS8602 warning — factory is guaranteed non-null here because
    the preceding try/catch would have returned on failure

Files changed:
  - Phases/Phase1_DeviceTest.cs (1 warning fix)
  - Phases/Phase3_TextureOwnership.cs (2 uint->int casts)
  - Phases/Phase4_NVENCRegistration.cs (2 uint->int casts)
  - Phases/Phase5_PerformanceBenchmark.cs (2 int->uint casts)

No changes to Foundation, Contract, or production backends.

Authored by: GLM-2 (GPU Pipeline Validation Engineer)
Task ID: P1-B.2-V1-Spike (build fix #5)

+++ spikes/D3D11_NVENC_Spike/Phases/Phase3_TextureOwnership.cs
@@ -83,8 +83,10 @@ public static class Phase3_TextureOwnership
         // --- Step 3: Verify dimensions match desktop ---
         Console.WriteLine();
         Console.WriteLine("[3.3] Verifying dimensions match desktop...");
-        int desktopWidth = duplDesc.ModeDescription.Width;
-        int desktopHeight = duplDesc.ModeDescription.Height;
+        // Vortice's ModeDescription.Width/Height are uint — cast to int for
+        // comparison with desc.Width (which is int in Texture2DDescription).
+        int desktopWidth = (int)duplDesc.ModeDescription.Width;
+        int desktopHeight = (int)duplDesc.ModeDescription.Height;
         Console.WriteLine($"  Desktop:  {desktopWidth}x{desktopHeight}");
         Console.WriteLine($"  Texture:  {desc.Width}x{desc.Height}");
         if (desc.Width != desktopWidth || desc.Height != desktopHeight)
